Work on adding new orthophoto sources
