+++ AeroScenery/AeroSceneryManager.cs
@@ -69,7 +69,7 @@ namespace AeroScenery
             dataRepository = new SqlLiteDataRepository();
 
             imageTiles = null;
-            version = "1.0.2";
+            version = "1.1.0";
             incrementalVersion = 10;
         }
