+++ AeroScenery/changelog.txt
@@ -11,6 +11,11 @@
 
 Change log
 
+1.1
+- Added ArcGIS orthophoto support
+- Fixed issue where 'Shrink TMC grid squares" used culture specific decimal separator
+
+
 1.0.2
 
 - Fixed potential issue when hex folder names are changed
